Include a default slack notification template (#9)
* Include a default slack notification template

This fixes https://github.com/simplifi/terraform-google-cloud-build-slack-notifier/issues/7

Prior to this change, the notifier cloud run invocation would fail because no template for the Slack message could be found. This PR will ensure that a template is made available. By default it uses the slack.json file added in this PR, but this can be overridden with the new override_slack_template_json input variable. If this is provided, it should be the contents of the template, which will be stored in the GCS bucket and used instead of the default.

slack.json is based on https://github.com/GoogleCloudPlatform/cloud-build-notifiers/blob/master/slack/slack.json but with the build status fixed.

* Fixed the Travis checks